Revert "Segregated-filter expression-based item search"
