+++ .gitignore
@@ -1,6 +1,3 @@
 *.swp
 *.bat
 sets/
-
-.idea/
-*.iml
