+++ builder.js
@@ -3,18 +3,11 @@ const url_tag = location.hash.slice(1);
 console.log(url_base);
 console.log(url_tag);
 
-const BUILD_VERSION = "6.9.9";
+const BUILD_VERSION = "6.9.7";
 
 function setTitle() {
-    let text;
-    if (url_base.includes("hppeng-wynn")) {
-        text = "WynnBuilder UNSTABLE version "+BUILD_VERSION+" (db version "+DB_VERSION+")";
-    }
-    else {
-        text = "WynnBuilder version "+BUILD_VERSION+" (db version "+DB_VERSION+")";
-        document.getElementById("header").classList.add("funnynumber");
-    }
-    document.getElementById("header").textContent = text;
+    document.getElementById("header").textContent = "WynnBuilder version "+BUILD_VERSION+" (db version "+DB_VERSION+")";
+    document.getElementById("header").classList.add("funnynumber");
 }
 
 setTitle();
@@ -78,12 +71,63 @@ let equipment_names = [
 let equipmentInputs = equipment_fields.map(x => x + "-choice");
 let buildFields = equipment_fields.map(x => "build-"+x);
 
+let powderIDs = new Map();
+let powderNames = new Map();
+let _powderID = 0;
+for (const x of skp_elements) {
+    for (let i = 1; i <= 6; ++i) {
+        // Support both upper and lowercase, I guess.
+        powderIDs.set(x.toUpperCase()+i, _powderID);
+        powderIDs.set(x+i, _powderID);
+        powderNames.set(_powderID, x+i);
+        _powderID++;
+    }
+}
 let powderInputs = [
     "helmet-powder",
     "chestplate-powder",
     "leggings-powder",
     "boots-powder",
     "weapon-powder",
+];
+// Ordering: [dmgMin, dmgMax, convert, defPlus, defMinus (+6 mod 5)]
+class Powder {
+    constructor(min, max, convert, defPlus, defMinus) {
+        this.min = min;
+        this.max = max;
+        this.convert = convert;
+        this.defPlus = defPlus;
+        this.defMinus = defMinus;
+    }
+}
+function _p(a,b,c,d,e) { return new Powder(a,b,c,d,e); } //bruh moment
+
+let powderStats = [
+    _p(3,6,17,2,1), _p(6,9,21,4,2), _p(8,14,25,8,3), _p(11,16,31,14,5), _p(15,18,38,22,9), _p(18,22,46,30,13),
+    _p(1,8,9,3,1), _p(1,13,11,5,1), _p(2,18,14,9,2), _p(3,24,17,14,4), _p(3,32,22,20,7), _p(5,40,28,28,10),
+    _p(3,4,13,3,1), _p(4,7,15,6,1), _p(6,10,17,11,2), _p(8,12,21,18,4), _p(11,14,26,28,7), _p(13,17,32,40,10),
+    _p(2,5,14,3,1), _p(4,8,16,5,2), _p(6,10,19,9,3), _p(9,13,24,16,5), _p(12,16,30,25,9), _p(15,19,37,36,13),
+    _p(2,6,11,3,1), _p(4,9,14,6,2), _p(7,10,17,10,3), _p(9,13,22,16,5), _p(13,18,28,24,9), _p(16,18,35,34,13)
+];
+
+//Ordering: [weapon special name, weapon special effects, armor special name, armor special effects]
+class PowderSpecial{
+    constructor(wSpName, wSpEff, aSpName, aSpEff, cap){
+        this.weaponSpecialName = wSpName;
+        this.weaponSpecialEffects = wSpEff;
+        this.armorSpecialName = aSpName;
+        this.armorSpecialEffects = aSpEff;
+        this.cap = cap;
+    }
+}
+function _ps(a,b,c,d,e) { return new PowderSpecial(a,b,c,d,e); } //bruh moment
+
+let powderSpecialStats = [
+    _ps("Quake",new Map([["Radius",[5,5.5,6,6.5,7]], ["Damage",[155,220,285,350,415]] ]),"Rage",new Map([ ["Damage", [0.3,0.4,0.5,0.7,1.0]],["Description", "% " + "\u2764" + " Missing"] ]),400), //e
+    _ps("Chain Lightning",new Map([ ["Chains", [5,6,7,8,9]], ["Damage", [200,225,250,275,300]] ]),"Kill Streak",new Map([ ["Damage", [3,4.5,6,7.5,9]],["Duration", [5,5,5,5,5]],["Description", "Mob Killed"] ]),200), //t
+    _ps("Curse",new Map([ ["Duration", [7,7.5,8,8.5,9]],["Damage Boost", [90,120,150,180,210]] ]),"Concentration",new Map([ ["Damage", [1,2,3,4,5]],["Duration",[1,1,1,1,1]],["Description", "Mana Used"] ]),150), //w
+    _ps("Courage",new Map([ ["Duration", [6,6.5,7,7.5,8]],["Damage", [75,87.5,100,112.5,125]],["Damage Boost", [70,90,110,130,150]] ]),"Endurance",new Map([ ["Damage", [2,3,4,5,6]],["Duration", [8,8,8,8,8]],["Description", "Hit Taken"] ]),200), //f
+    _ps("Air Prison",new Map([ ["Duration", [3,3.5,4,4.5,5]],["Damage Boost", [400,450,500,550,600]],["Knockback", [8,12,16,20,24]] ]),"Dodge",new Map([ ["Damage",[2,3,4,5,6]],["Duration",[2,3,4,5,6]],["Description","Near Mobs"] ]),150) //a
 ];
 
 let itemTypes = armorTypes.concat(accessoryTypes).concat(weaponTypes);
@@ -94,7 +138,6 @@ for (const it of itemTypes) {
 let itemMap = new Map();
 /* Mapping from item names to set names. */
 let idMap = new Map();
-let redirectMap = new Map();
 
 /*
  * Function that takes an item list and populates its corresponding dropdown.
@@ -151,18 +194,13 @@ function init() {
     items = items.concat(noneItems);
     console.log(items);
     for (const item of items) {
-        if (item.remapID === undefined) {
-            itemLists.get(item.type).push(item.displayName);
-            itemMap.set(item.displayName, item);
-            if (noneItems.includes(item)) {
-                idMap.set(item.id, "");
-            }
-            else {
-                idMap.set(item.id, item.displayName);
-            }
+        itemLists.get(item.type).push(item.displayName);
+        itemMap.set(item.displayName, item);
+        if (noneItems.includes(item)) {
+            idMap.set(item.id, "");
         }
         else {
-            redirectMap.set(item.id, item.remapID);
+            idMap.set(item.id, item.displayName);
         }
     }
     
@@ -217,13 +255,6 @@ function init() {
     decodeBuild(url_tag);
 }
 
-function getItemNameFromID(id) {
-    if (redirectMap.has(id)) {
-        return getItemNameFromID(redirectMap.get(id));
-    }
-    return idMap.get(id);
-}
-
 /*
  * Populate fields based on url, and calculate build.
  */
@@ -239,7 +270,7 @@ function decodeBuild(url_tag) {
         if (version === "0" || version === "1" || version === "2" || version === "3") {
             let equipments = info[1];
             for (let i = 0; i < 9; ++i ) {
-                equipment[i] = getItemNameFromID(Base64.toInt(equipments.slice(i*3,i*3+3)));
+                equipment[i] = idMap.get(Base64.toInt(equipments.slice(i*3,i*3+3)));
             }
         }
         if (version === "1") {
@@ -736,8 +767,7 @@ function calculateBuildStats() {
     //skpRow.classList.add("left");
     let td = document.createElement("p");
     //td.classList.add("left");
-
-    /* let skpSummary = document.createElement("b");
+    let skpSummary = document.createElement("b");
     skpSummary.textContent = "Assigned " + player_build.assigned_skillpoints + " skillpoints. Total: (";
     //skpSummary.classList.add("itemp");
     td.appendChild(skpSummary);
@@ -756,12 +786,12 @@ function calculateBuildStats() {
     let skpEnd = document.createElement("b");
     skpEnd.textContent = ")";
     td.appendChild(skpEnd);
-    skpRow.append(td); */
+    skpRow.append(td);
 
     let remainingSkp = document.createElement("p");
     remainingSkp.classList.add("center");
     let remainingSkpTitle = document.createElement("b");
-    remainingSkpTitle.textContent = "Assigned " + player_build.assigned_skillpoints + " skillpoints. Remaining skillpoints: ";
+    remainingSkpTitle.textContent = "Remaining skillpoints: ";
     let remainingSkpContent = document.createElement("b");
     //remainingSkpContent.textContent = "" + (levelToSkillPoints(player_build.level) - player_build.assigned_skillpoints < 0 ? "< 0" : levelToSkillPoints(player_build.level) - player_build.assigned_skillpoints);
     remainingSkpContent.textContent = "" + (levelToSkillPoints(player_build.level) - player_build.assigned_skillpoints);
