+++ crafter.html
@@ -100,26 +100,6 @@
                         </td>
                     </tr>
                 </table>
-                <label for = "atkSpdChoices"> Attack Speed (weapons only):</label>
-                <table class = "center" id = "atkSpdChoices">
-                    <tr>
-                        <td class = "center">
-                            <button class = "button" id = "slow-atk-button" onclick = "toggleAtkSpd('slow-atk-button')">
-                            Slow
-                            </button>
-                        </td>
-                        <td class = "center">
-                            <button class = "button" id = "normal-atk-button" onclick = "toggleAtkSpd('normal-atk-button')">
-                            Normal
-                            </button>
-                        </td>
-                        <td class = "center">
-                            <button class = "button" id = "fast-atk-button" onclick = "toggleAtkSpd('fast-atk-button')">
-                            Fast
-                            </button>
-                        </td>
-                    </tr>
-                </table>
                 <table class = "center">
                     <tr>
                         <td class = "center">
@@ -145,12 +125,12 @@
                     </tr>
                 </table>
             </div>
-            <div class = "recipe center hide-container-block" style = "display:none">
+            <div class = "recipe hide-container-block" style = "display:none">
                 <div class = "recipe-stats">
                     <div class = "center" id = "recipe-stats"></div>
                 </div>
             </div>
-            <div class = "crafted center hide-container-block" style = "display:none">
+            <div class = "crafted hide-container-block" style = "display:none">
                 <div class = "craft-stats">
                     <div class = "center" id = "craft-stats"></div>
                 </div>
@@ -160,8 +140,10 @@
             <div class = "craft-warnings">
                 <div class = "center" id = "craft-warnings"></div>
             </div>
-            <p></p>
-            <p></p>
+            <p>
+            </p>
+            <p>
+            </p>
             <div class="ingredients-container hide-container-grid" id = "ingreds" style = "display:none">
                 <p class="center title hide-container-block" style = "display:block">
                     Ingredients
@@ -198,7 +180,6 @@
         </div>
 
         <script type="text/javascript" src="utils.js"></script>
-        <script type="text/javascript" src="powders.js"></script>
         <script type="text/javascript" src="build_utils.js"></script>
         <script type="text/javascript" src="skillpoints.js"></script>
         <script type="text/javascript" src="damage_calc.js"></script>
