+++ crafter.js
@@ -86,18 +86,6 @@ function updateMaterials() {
         document.getElementById("mat-2").textContent = "Material 2 Tier:";
     }
 }
-function toggleAtkSpd(buttonId) {
-    let buttons = ["slow-atk-button", "normal-atk-button", "fast-atk-button"];
-    let elem = document.getElementById(buttonId);
-    if (elem.classList.contains("toggleOn")) {
-        elem.classList.remove("toggleOn");
-    } else {
-        for (const button of buttons) {
-            document.getElementById(button).classList.remove("toggleOn");
-        }
-        elem.classList.add("toggleOn");
-    }
-}
 
 function calculateCraft() {
     //Make things display.
@@ -129,15 +117,9 @@ function calculateCraft() {
     for (i = 1; i < 7; i++) {
         getValue("ing-choice-" + i) === "" ? ingreds.push(expandIngredient(ingMap.get("No Ingredient"))) : ingreds.push(expandIngredient(ingMap.get(getValue("ing-choice-" + i))));
     }
-    let atkSpd = "NORMAL"; //default attack speed will be normal.
-    for (const b of ["slow-atk-button", "normal-atk-button", "fast-atk-button"]) {
-        button = document.getElementById(b);
-        if (button.classList.contains("toggleOn")) {
-            atkSpd = b.split("-")[0].toUpperCase();
-        }
-    }
+
     //create the craft
-    player_craft = new Craft(recipe,mat_tiers,ingreds,atkSpd);
+    player_craft = new Craft(recipe,mat_tiers,ingreds);
     console.log(player_craft);
     /*console.log(recipe)
     console.log(levelrange)
@@ -148,9 +130,6 @@ function calculateCraft() {
     
     //Display Recipe Stats
     displayRecipeStats(player_craft, "recipe-stats");
-    for(let i = 0; i < 6; i++) {
-        displayExpandedIngredient(player_craft["ingreds"][i],"tooltip-" + i);
-    }
     //Display Craft Stats
     displayCraftStats(player_craft, "craft-stats");
     //Display Ingredients' Stats
